+++ tests/test_utils.py
@@ -1,5 +1,5 @@
 from pycfmodel.constants import IMPLEMENTED_FUNCTIONS
-from pycfmodel.model.intrinsic_function import FUNCTION_MAPPINGS
+from pycfmodel.resolver import FUNCTION_MAPPINGS
 
 
 def test_implemented_functions():
